[Feature] 성은의 변경사항
destination_select_screen.dart 파일의 Appbar에 마이페이지 버튼 추가함.
my_page.dart 파일을 구조적으로 정리하기 위해 my_page_content.dart와 my_page_popup.dart로 분리함.
